validator/requirements.txt
# Core dependencies
networkx>=3.0
pyyaml>=6.0

# Optional: for advanced metrics
numpy>=1.24
scipy>=1.10
